requirements.txt
# Core dependencies
torch>=1.9.0
torchaudio>=0.9.0
transformers>=4.20.0
openai-whisper>=20230117
faster-whisper>=0.7.0
numpy>=1.19.0
scipy>=1.5.0
librosa>=0.9.0

# Audio processing
pyaudio>=0.2.11
soundfile>=0.10.3
speechbrain>=0.5.12
demucs>=4.0.0

# NLP and ML
editdistance>=0.5.3
phonemizer>=3.0.0
nltk>=3.7
spacy>=3.4.0

# Real-time processing
websockets>=10.0
asyncio>=3.4.3
aiohttp>=3.8.0

# Utilities
pyyaml>=5.4
python-dotenv>=0.19.0
tqdm>=4.62.0
colorama>=0.4.4
loguru>=0.6.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.18.0
pytest-cov>=3.0.0

# Optional optimizations
onnxruntime>=1.10.0
tensorrt>=8.0.0  # For NVIDIA GPUs